backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database Drivers
psycopg2-binary==2.9.9
pymongo==4.6.1
mysql-connector-python==8.3.0
sqlalchemy==2.0.25

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Task Scheduling
apscheduler==3.10.4

# Monitoring & Metrics
psutil==5.9.8

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# CORS
python-cors==1.0.0

# Email (para alertas)
aiosmtplib==3.0.1

# HTTP Client
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
